[change-owners] split tests into topic related files
